+++ changelog.md
@@ -1,4 +1,11 @@
 ## 1.6.0
+### Added
+- Guideline available in More tab
+- App author information available in More tab
+
+### Improved
+- Header on all screens to be more consistent
+
 ### Fixed
 - Not saving language code